refactor: split monolith utils into focused files, move NAME_CHAR_REG and REMOTE_FROM_PARAMETER to constants
Amp-Thread-ID: https://ampcode.com/threads/T-019d3464-cde5-708b-9128-0bbe1c9a30cb
