+++ packages/lib/src/constants/remote.ts
@@ -1,0 +1,2 @@
+export const REMOTE_FROM_PARAMETER = 'remoteFrom';
+export const NAME_CHAR_REG = new RegExp('[0-9a-zA-Z@_-]+');
