+++ packages/lib/src/utils/array.ts
@@ -1,0 +1,3 @@
+export const uniqueArr = <T>(arr: T[]): T[] => {
+  return Array.from(new Set(arr));
+};
