+++ packages/lib/src/utils/index.ts
@@ -13,303 +13,16 @@
 // SPDX-License-Identifier: MulanPSL-2.0
 // *****************************************************************************
 
-import type {
-  ConfigTypeSet,
-  Exposes,
-  Remotes,
-  RemotesConfig,
-  Shared,
-  VitePluginFederationOptions
-} from '../../types';
-import { readFileSync } from 'node:fs';
-import { createHash } from 'node:crypto';
-import path, { parse, posix } from 'node:path';
-import type { ResolvedConfig, Rolldown } from 'vite';
-import type { ServerResponse } from 'node:http';
-
+export * from './array';
+export * from './dependency';
+export * from './file';
+export * from './hash';
 export * from './html';
-
-export const findDependencies = (
-  ctx: Rolldown.PluginContext,
-  id: string,
-  sets: Set<string>,
-  sharedModuleIds: Map<string, string>,
-  usedSharedModuleIds: Set<string>
-): void => {
-  if (!sets.has(id)) {
-    sets.add(id);
-    const moduleInfo = ctx.getModuleInfo(id);
-    if (moduleInfo?.importedIds) {
-      moduleInfo.importedIds.forEach((id) => {
-        findDependencies(ctx, id, sets, sharedModuleIds, usedSharedModuleIds);
-      });
-    }
-    if (sharedModuleIds.has(id)) {
-      usedSharedModuleIds.add(sharedModuleIds.get(id) as string);
-    }
-  }
-};
-
-export const parseSharedOptions = (
-  options: VitePluginFederationOptions
-): (string | ConfigTypeSet)[] => {
-  return parseOptions(
-    options.shared || {},
-    (value, key) => ({
-      import: true,
-      shareScope: 'default',
-      packagePath: key,
-      // Whether the path is set manually
-      manuallyPackagePathSetting: false,
-      generate: true,
-      modulePreload: false
-    }),
-    (value, key) => {
-      value.import = value.import ?? true;
-      value.shareScope = value.shareScope || 'default';
-      value.packagePath = value.packagePath || key;
-      value.manuallyPackagePathSetting = value.packagePath !== key;
-      value.generate = value.generate ?? true;
-      value.modulePreload = value.modulePreload ?? false;
-      return value;
-    }
-  );
-};
-
-export const parseExposeOptions = (
-  options: VitePluginFederationOptions
-): (string | ConfigTypeSet)[] => {
-  return parseOptions(
-    options.exposes,
-    (item) => {
-      return {
-        import: item,
-        name: undefined,
-        dontAppendStylesToHead: false
-      };
-    },
-    (item) => ({
-      import: item.import,
-      name: item.name || undefined,
-      dontAppendStylesToHead: item.dontAppendStylesToHead || false
-    })
-  );
-};
-
-export const createContentHash = (path: string): string => {
-  const content = readFileSync(path, { encoding: 'utf-8' });
-  return createHash('md5').update(content).digest('hex').toString().slice(0, 8);
-};
-
-export const parseRemoteOptions = (
-  options: VitePluginFederationOptions
-): (string | ConfigTypeSet)[] => {
-  return parseOptions(
-    options.remotes ? options.remotes : {},
-    (item) => ({
-      external: Array.isArray(item) ? item : [item],
-      shareScope: options.shareScope || 'default',
-      format: 'esm',
-      from: 'vite',
-      externalType: 'url'
-    }),
-    (item) => ({
-      external: Array.isArray(item.external) ? item.external : [item.external],
-      shareScope: item.shareScope || options.shareScope || 'default',
-      format: item.format || 'esm',
-      from: item.from ?? 'vite',
-      externalType: item.externalType || 'url'
-    })
-  );
-};
-
-export const parseOptions = (
-  options: Exposes | Remotes | Shared | undefined,
-  normalizeSimple: (value: any, key: any) => ConfigTypeSet,
-  normalizeOptions: (value: any, key: any) => ConfigTypeSet
-): (string | ConfigTypeSet)[] => {
-  if (!options) {
-    return [];
-  }
-  const list: {
-    [index: number]: string | ConfigTypeSet;
-  }[] = [];
-  const array = (items: (string | ConfigTypeSet)[]) => {
-    for (const item of items) {
-      if (typeof item === 'string') {
-        list.push([item, normalizeSimple(item, item)]);
-      } else if (item && typeof item === 'object') {
-        object(item);
-      } else {
-        throw new Error('Unexpected options format');
-      }
-    }
-  };
-  const object = (obj) => {
-    for (const [key, value] of Object.entries(obj)) {
-      if (typeof value === 'string' || Array.isArray(value)) {
-        list.push([key, normalizeSimple(value, key)]);
-      } else {
-        list.push([key, normalizeOptions(value, key)]);
-      }
-    }
-  };
-  if (Array.isArray(options)) {
-    array(options);
-  } else if (typeof options === 'object') {
-    object(options);
-  } else {
-    throw new Error('Unexpected options format');
-  }
-  return list;
-};
-
-const letterReg = new RegExp('[0-9a-zA-Z]+');
-
-export const removeNonRegLetter = (str: string, reg = letterReg): string => {
-  let needUpperCase = false;
-  let ret = '';
-  for (const c of str) {
-    if (reg.test(c)) {
-      ret += needUpperCase ? c.toUpperCase() : c;
-      needUpperCase = false;
-    } else {
-      needUpperCase = true;
-    }
-  }
-  return ret;
-};
-
-export const getModuleMarker = (value: string, type?: string): string => {
-  return type ? `__rf_${type}__${value}` : `__rf_placeholder__${value}`;
-};
-
-export const normalizePath = (id: string): string => {
-  return posix.normalize(id.replace(/\\/g, '/'));
-};
-
-export const uniqueArr = <T>(arr: T[]): T[] => {
-  return Array.from(new Set(arr));
-};
-
-export const isSameFilepath = (src: string, dest: string): boolean => {
-  if (!src || !dest) {
-    return false;
-  }
-  src = normalizePath(src);
-  dest = normalizePath(dest);
-  const srcExt = parse(src).ext;
-  const destExt = parse(dest).ext;
-  if (srcExt && destExt && srcExt !== destExt) {
-    return false;
-  }
-  if (srcExt) {
-    src = src.slice(0, -srcExt.length);
-  }
-  if (destExt) {
-    dest = dest.slice(0, -destExt.length);
-  }
-  return src === dest;
-};
-
-export type Remote = { id: string; regexp: RegExp; config: RemotesConfig };
-
-export const createRemotesMap = (remotes: Remote[]): string => {
-  const createUrl = (remote: Remote) => {
-    const external = remote.config.external[0];
-    const externalType = remote.config.externalType;
-    if (externalType === 'promise') {
-      return `()=>${external}`;
-    } else {
-      return `'${external}'`;
-    }
-  };
-  return `const remotesMap = {
-${remotes
-  .map(
-    (remote) =>
-      `'${remote.id}':{url:${createUrl(remote)},format:'${
-        remote.config.format
-      }',from:'${remote.config.from}'}`
-  )
-  .join(',\n  ')}
-};`;
-};
-
-/**
- * get file extname from url
- * @param url
- */
-export const getFileExtname = (url: string): string => {
-  const fileNameAndParamArr = normalizePath(url).split('/');
-  const fileNameAndParam = fileNameAndParamArr[fileNameAndParamArr.length - 1];
-  const fileName = fileNameAndParam.split('?')[0];
-  return path.extname(fileName);
-};
-
-export const REMOTE_FROM_PARAMETER = 'remoteFrom';
-export const NAME_CHAR_REG = new RegExp('[0-9a-zA-Z@_-]+');
-
-/** Serialize an array of strings into a JS array literal (e.g. `['a','b']`). */
-export const toJsArrayLiteral = (items: string[]): string =>
-  `[${items.map((s) => JSON.stringify(s)).join(',')}]`;
-
-export const joinUrlSegments = (a: string, b: string): string => {
-  if (!a || !b) {
-    return a || b || '';
-  }
-  if (a[a.length - 1] === '/') {
-    a = a.substring(0, a.length - 1);
-  }
-  if (b[0] !== '/') {
-    b = '/' + b;
-  }
-  return a + b;
-};
-
-export const toOutputFilePathWithoutRuntime = (
-  filename: string,
-  type: 'asset' | 'public',
-  hostId: string,
-  hostType: 'js' | 'css' | 'html',
-  config: ResolvedConfig,
-  toRelative: (filename: string, hostId: string) => string
-): string => {
-  const { renderBuiltUrl } = config.experimental;
-  let relative = config.base === '' || config.base === './';
-  if (renderBuiltUrl) {
-    const result = renderBuiltUrl(filename, {
-      hostId,
-      hostType,
-      type,
-      ssr: !!config.build.ssr
-    });
-    if (typeof result === 'object') {
-      if (result.runtime) {
-        throw new Error(
-          `{ runtime: "${result.runtime}" } is not supported for assets in ${hostType} files: ${filename}`
-        );
-      }
-      if (typeof result.relative === 'boolean') {
-        relative = result.relative;
-      }
-    } else if (result) {
-      return result;
-    }
-  }
-  if (relative && !config.build.ssr) {
-    return toRelative(filename, hostId);
-  } else {
-    return joinUrlSegments(config.base, filename);
-  }
-};
-
-/** Check whether a request URL matches a path (with or without query string). */
-export const matchesUrl = (url: string | undefined, path: string): boolean =>
-  url === path || !!url?.startsWith(`${path}?`);
-
-/** Send a JavaScript response. */
-export const sendJs = (res: ServerResponse, code: string): void => {
-  res.setHeader('Content-Type', 'application/javascript');
-  res.end(code);
-};
+export * from './http';
+export * from './literal';
+export * from './module';
+export * from './package';
+export * from './plugin';
+export * from './remote';
+export * from './string';
+export * from './url';
