+++ packages/lib/src/utils/literal.ts
@@ -1,0 +1,2 @@
+export const toJsArrayLiteral = (items: string[]): string =>
+  `[${items.map((s) => JSON.stringify(s)).join(',')}]`;
