+++ packages/lib/src/utils/module.ts
@@ -1,0 +1,3 @@
+export const getModuleMarker = (value: string, type?: string): string => {
+  return type ? `__rf_${type}__${value}` : `__rf_placeholder__${value}`;
+};
